requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
<<<<<<< HEAD
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord==2.1.0
discord.py==2.1.0
Django==4.1.5
django-filter==22.1
django-jsonview==2.0.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
multidict==6.0.4
openai==0.26.0
psycopg2-binary==2.9.5
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sodapy==2.2.0
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13
yarl==1.8.2
youtube-dl==2021.12.17
=======
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
discord==2.1.0
discord.py==2.1.0
Django==4.1.3
django-filter==22.1
django-jsonview==2.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.1.0
Markdown==3.4.1
multidict==6.0.4
mysql-connector-python==8.0.31
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.230105
protobuf==3.20.1
psycopg2-binary==2.9.5
pycparser==2.21
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sodapy==2.2.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.7.0.0
typing-extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
>>>>>>> 0ec535af554e4234438853d1b8009bf6d88a0304
zipp==3.11.0
